requirements.txt
boto3>=1.26.0
botocore>=1.29.0
python-dotenv>=1.0.0

pytest>=7.0.0
coverage>=6.0.0
